+++ backend/src/workers/soroban-event-worker.ts
@@ -50,9 +50,9 @@ export function decodeAddress(val: xdr.ScVal): string {
   ) {
     return StrKey.encodeEd25519PublicKey(addr.accountId().ed25519());
   }
-  // addr.contractId() returns a Hash (Opaque[]); cast to Uint8Array for encodeContract
+  // addr.contractId() returns a Hash (Opaque[])
   const hash = addr.contractId();
-  return StrKey.encodeContract(Buffer.from(hash as Uint8Array));
+  return StrKey.encodeContract(Buffer.from(hash));
 }
 
 /**
